refactor: features/profile 배럴 패턴 적용
