+++ src/features/profile/image/model/index.ts
@@ -1,0 +1,1 @@
+export { default as useUploadProfileImageMutation } from './use-upload-profile-image'
