+++ src/features/profile/password-change/model/index.ts
@@ -1,0 +1,1 @@
+export { default as usePasswordChangeMutation } from './use-password-change-mutation'
