+++ src/features/profile/password-change/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as PasswordChangeForm } from './password-change-form'
